separate browse and search tests
